requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
Flask-Mail
FLASK-CORS
python-dateutil